requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.3
Brotli==1.1.0
requests
beautifulsoup4
python-telegram-bot
python-dotenv
phonenumbers